Register @ConfigMappings directly into the Config builder
