+++ core/runtime/src/main/java/io/quarkus/runtime/configuration/ConfigUtils.java
@@ -191,6 +191,15 @@ public final class ConfigUtils {
         builder.withSources(provider.getConfigSourceFactory(Thread.currentThread().getContextClassLoader()));
     }
 
+    public static void addMapping(SmallRyeConfigBuilder builder, String mappingClass, String prefix) {
+        ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
+        try {
+            builder.withMapping(contextClassLoader.loadClass(mappingClass), prefix);
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     /**
      * Checks if a property is present in the current Configuration.
      *
