+++ integration-tests/smallrye-config/src/main/java/io/quarkus/it/smallrye/config/ExceptionConfig.java
@@ -1,0 +1,8 @@
+package io.quarkus.it.smallrye.config;
+
+import io.smallrye.config.ConfigMapping;
+
+@ConfigMapping(prefix = "exception")
+public interface ExceptionConfig {
+    String message();
+}
